# openapi.yaml
openapi: 3.0.3
info:
  title: BookedIRL Backend API
  version: 0.1.1
servers:
  - url: http://localhost:4000
  - url: https://bookedirlbackend.onrender.com
...
paths:
...
  /users/me:
    get:
      tags: [users]
      security:
        - bearerAuth: []
      summary: Get current user profile
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found
  /users:
    post:
      tags: [users]
      security:
        - bearerAuth: []
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "409":
          description: Username taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{userId}:
    get:
      tags: [users]
      security:
        - bearerAuth: []
      summary: Get user profile by id
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found
  /users/nearby:
    get:
      tags: [users]
      security:
        - bearerAuth: []
      summary: Nearby users by location
      parameters:
        - $ref: "#/components/parameters/Lat"
        - $ref: "#/components/parameters/Lng"
        - $ref: "#/components/parameters/RadiusKm"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NearbyUser"
  /users/nearby-qualified:
    get:
      tags: [users]
      security:
        - bearerAuth: []
      summary: Nearby users with photos and active services
      parameters:
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NearbyUser"
  /users/photos:
    get:
      tags: [users]
      security:
        - bearerAuth: []
      summary: Photos for a list of users
      parameters:
        - name: userIds
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserPhotoItem"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Lat:
      name: lat
      in: query
      required: true
      schema:
        type: number
        format: double
    Lng:
      name: lng
      in: query
      required: true
      schema:
        type: number
        format: double
    RadiusKm:
      name: radiusKm
      in: query
      required: false
      schema:
        type: number
        default: 25
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        default: 20
...
  schemas:
...
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
      required: [ok]
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    UserProfile:
      type: object
      properties:
        id:
          type: string
        display_name:
          type: string
          nullable: true
        username:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        intent_looking:
          type: boolean
        intent_offering:
          type: boolean
        photos:
          type: array
          items:
            $ref: "#/components/schemas/UserPhotoItem"
        socialLinks:
          type: array
          items:
            $ref: "#/components/schemas/UserSocialLink"
      required: [id, intent_looking, intent_offering, photos, socialLinks]
    UserUpdate:
      type: object
      properties:
        displayName:
          type: string
          nullable: true
        username:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        intentLooking:
          type: boolean
        intentOffering:
          type: boolean
        photos:
          type: array
          items:
            type: string
        socialLinks:
          type: array
          items:
            type: string
        location:
          type: object
          properties:
            lat:
              type: number
            lng:
              type: number
      additionalProperties: false
    UserPhotoItem:
      type: object
      properties:
        url:
          type: string
        sort_order:
          type: integer
      required: [url]
    UserSocialLink:
      type: object
      properties:
        url:
          type: string
        sort_order:
          type: integer
      required: [url]
    NearbyUser:
      type: object
      properties:
        id:
          type: string
        display_name:
          type: string
          nullable: true
        username:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        distance_meters:
          type: number
        photos:
          type: array
          items:
            $ref: "#/components/schemas/UserPhotoItem"
      required: [id, distance_meters, photos]
...